Update main.yaml
This error means the `curl` command-line tool, which the playbook is trying to use to download Consul, is not installed on the target machine.

The error message `[Errno 2] No such file or directory: b'curl'` is the key. The system tried to execute the `curl` command but couldn't find the program.

-----

### \#\# The Problem: Missing Dependency

Even though you have a task named `Install common packages`, it appears that **`curl` is not included in its list of packages to be installed**. The playbook then proceeds to the `consul` role and fails because it depends on `curl` to download the software.

-----

### \#\# The Solution: Add `curl` to Your Package List

You need to edit your `common` role to ensure `curl` is installed before any other roles try to use it.

1.  **Open the file** for the common packages task. Based on your playbook structure, this is likely: `/home/user/llama-cluster-upbringing-script/ansible/roles/common/tasks/main.yaml`.
2.  **Find the package list** and add `curl` to it.

#### **Example Ansible Task Fix**

Your `Install common packages` task probably looks something like this:

**Before:**

```yaml
- name: Install common packages
  ansible.builtin.package:
    name:
      - vim
      - git
      - chrony
    state: present
```

**After:**

```yaml
- name: Install common packages
  ansible.builtin.package:
    name:
      - vim
      - git
      - chrony
      - curl   # <-- Add this line
    state: present
```

By adding `curl` to this list, you ensure it gets installed early in the playbook run. When the `Download Consul` task is executed later, the `curl` command will be available, and the download will succeed. ✅