Update main.yaml
The Problem: Mismatched Version and Filename
By comparing your Ansible code with the new release information, we can see a few issues:

Version: Your playbook is set to v1.2.1-rc1, but the latest version is v2.1.0.

Filename: The new release files are named paddler-bin-linux-libc..., not the paddler-{{ paddler_version }}-linux-amd64.tar.gz format your playbook is trying to create.

File Type: The new assets are binary executables, not .tar.gz archives. This means the unarchive step in your playbook is incorrect and needs to be removed.